Fix pickle issue if 'spawn' or 'forkserver' multiprocessing start method
When activating python multiprocessing.set_start_method to 'forkserver' or
'spawn', a Pickle exceptions was raised.
('forkserver' was needed for me to lower memory usage)

So I modified Process cretation in OSMParser
so that its start function and parameters dont reference
the OSMParser instance itself or the callbacks.

Then the OSMParser and the classes containing the callbacks method
will not need to be pickled and reinstantiated in the created parser process.